rag-benchmark/requirements.txt
google-adk
google-genai
faiss-cpu
pypdf
streamlit
python-dotenv
numpy
pypdf
rank-bm25